+++ sdk/signersdk.go
@@ -1,47 +1,47 @@
 package sdk
 
 import (
+	"encoding/json"
 	"errors"
 
+	"github.com/Golang-Tools/grpcsdk"
 	"github.com/Golang-Tools/jwthelper/v2/jwt_pb"
 	"github.com/Golang-Tools/jwthelper/v2/jwtsigner_pb"
 	"github.com/Golang-Tools/jwthelper/v2/signoptions"
-	"google.golang.org/grpc"
+	"github.com/Golang-Tools/loggerhelper/v2"
+	"github.com/Golang-Tools/optparams"
 )
 
-//NewSigner 建立一个新的签名器
-func (c *SDK) NewSigner() (*Signer, error) {
-	conn, err := newSigner(c, c.addr, c.opts...)
-	if err != nil {
-		return nil, err
-	}
-	return conn, nil
+type SignerSDK struct {
+	client *grpcsdk.SDK[jwtsigner_pb.JwtsignerClient]
 }
 
-//Signer 客户端类
-type Signer struct {
-	rpc  jwtsigner_pb.JwtsignerClient
-	conn *grpc.ClientConn
-	sdk  *SDK
+func NewSignerSDK() *SignerSDK {
+	s := new(SignerSDK)
+	s.client = grpcsdk.New(jwtsigner_pb.NewJwtsignerClient, &jwtsigner_pb.Jwtsigner_ServiceDesc)
+	return s
 }
 
-func newSigner(sdk *SDK, addr string, opts ...grpc.DialOption) (*Signer, error) {
-	c := new(Signer)
-	conn, err := grpc.Dial(addr, opts...)
-	if err != nil {
-		return nil, err
-	}
-	c.sdk = sdk
-	c.conn = conn
-	c.rpc = jwtsigner_pb.NewJwtsignerClient(conn)
-	return c, nil
+func (s *SignerSDK) Init(opts ...optparams.Option[grpcsdk.SDKConfig]) {
+	s.client.Init(opts...)
+}
+
+func (s *SignerSDK) GetLogger() *loggerhelper.Log {
+	return s.client.Logger
+}
+
+//Close 断开连接
+func (c *SignerSDK) Close() error {
+	return c.client.Close()
 }
 
 //Meta 查看远端签名器的元信息
-func (c *Signer) Meta() (*jwt_pb.SignerMeta, error) {
-	ctx, cancel := c.sdk.NewCtx()
+func (c *SignerSDK) Meta() (*jwt_pb.SignerMeta, error) {
+	ctx, cancel := c.client.NewCtx()
 	defer cancel()
-	res, err := c.rpc.Meta(ctx, &jwtsigner_pb.MetaRequest{})
+	Conn, release := c.client.GetClient()
+	defer release()
+	res, err := Conn.Meta(ctx, &jwtsigner_pb.MetaRequest{})
 	if err != nil {
 		return nil, err
 	}
@@ -58,7 +58,7 @@ func (c *Signer) Meta() (*jwt_pb.SignerMeta, error) {
 }
 
 // Sign 签名一个token
-func (c *Signer) Sign(payload interface{}, opts ...signoptions.SignOption) (*jwt_pb.Token, error) {
+func (c *SignerSDK) Sign(payload interface{}, opts ...optparams.Option[signoptions.SignOptions]) (*jwt_pb.Token, error) {
 	if payload == nil {
 		payload = map[string]interface{}{}
 	}
@@ -67,9 +67,7 @@ func (c *Signer) Sign(payload interface{}, opts ...signoptions.SignOption) (*jwt
 		return nil, err
 	}
 	defaultopt := signoptions.DefaultSignOptions
-	for _, opt := range opts {
-		opt.Apply(&defaultopt)
-	}
+	optparams.GetOption(&defaultopt, opts...)
 	query := jwtsigner_pb.SignRequest{
 		Sub:        defaultopt.Sub,
 		Exp:        defaultopt.Exp,
@@ -79,9 +77,11 @@ func (c *Signer) Sign(payload interface{}, opts ...signoptions.SignOption) (*jwt
 		Jti:        defaultopt.Jti,
 		Aud:        defaultopt.Aud,
 	}
-	ctx, cancel := c.sdk.NewCtx()
+	ctx, cancel := c.client.NewCtx()
 	defer cancel()
-	res, err := c.rpc.Sign(ctx, &query)
+	Conn, release := c.client.GetClient()
+	defer release()
+	res, err := Conn.Sign(ctx, &query)
 	if err != nil {
 		return nil, err
 	}
@@ -97,7 +97,8 @@ func (c *Signer) Sign(payload interface{}, opts ...signoptions.SignOption) (*jwt
 	return res.Token, nil
 }
 
-//Close 断开连接
-func (c *Signer) Close() error {
-	return c.conn.Close()
+var DefaultSigner *SignerSDK
+
+func init() {
+	DefaultSigner = NewSignerSDK()
 }