+++ sdk_bak/sdk_test.go
@@ -1,7 +1,0 @@
-package sdk
-
-type testPayLoad struct {
-	A int     `json:"a"`
-	B string  `json:"b,omitempty"`
-	C float32 `json:"-"`
-}