+++ universal.go
@@ -5,6 +5,7 @@ import (
 	"github.com/Golang-Tools/jwthelper/v2/jwt_pb"
 	"github.com/Golang-Tools/jwthelper/v2/signoptions"
 	"github.com/Golang-Tools/jwthelper/v2/verifyoptions"
+	"github.com/Golang-Tools/optparams"
 )
 
 //UniversalJwtSigner 通用jwt的签名器
@@ -12,7 +13,7 @@ type UniversalJwtSigner interface {
 	//Meta 查看签名器元信息
 	Meta() (*jwt_pb.SignerMeta, error)
 	// Sign 签名一个token
-	Sign(payload interface{}, opts ...signoptions.SignOption) (*jwt_pb.Token, error)
+	Sign(payload interface{}, opts ...optparams.Option[signoptions.SignOptions]) (*jwt_pb.Token, error)
 }
 
 //UniversalJwtVerifier 通用jwt的签名器
@@ -20,5 +21,5 @@ type UniversalJwtVerifier interface {
 	//Meta 查看签名器元信息
 	Meta() (*jwt_pb.VerifierMeta, error)
 	// 校验一个签名是否复合
-	Verify(token *jwt_pb.Token, payload interface{}, opts ...verifyoptions.VerifyOption) (*jwt_pb.JwtStatus, error)
+	Verify(token *jwt_pb.Token, payload interface{}, opts ...optparams.Option[verifyoptions.VerifyOptions]) (*jwt_pb.JwtStatus, error)
 }